Set secure_proxy_ssl_header 'HTTP_X_FORWARDED_PROTO' for nova-api
[mitaka-only] On master I3918f24c0c87cb626a28645b46e3df6360d5f924
makes it so nova uses the oslo middleware for the header parsing.

That isn't available in mitaka, so we can set the
secure_proxy_ssl_header explicitly via hieradata. Note this is
added to mitaka puppet-nova as an option in change:
I22deb886706fe71115a04fb52a7051be4783a5c4

Closes-Bug: 1606863

Change-Id: I9a5284869d3fca9615d7cbb84705a9cddd0df44a