Make deployed-server OS::Neutron::Port optional
This patch makes it possible to set
OS::TripleO::DeployedServer::ControlPlanePort: OS::Heat::None
in your resource_registry and thereby avoid the creation of
a neutron port for the deployed server. This is useful if
you are bootstrapping things in an environment without
Neutron.

Also, includes a new deployed-server-noop-ctlplane.yaml
environment file.

Change-Id: I2990dc816698e0f6e3193a8fc7c9c6767c6e50e5