Add hieradata for Manila v1 and v2 endpoints
manila::keystone::auth assumes that two endpoints will be created;
one for each version of Manila. This is not consistent with what we
provide to the manifest. So we then add the needed endpoint to the
EndpointMap.

v2 is left as the default (not needing to specify a version when
getting the value from the endpoint map), and v1 was added as an
extra suffix.

Change-Id: Ie8754fbb21f50c2613a2c05a52d771f4d1aeb4bb