Add parameter for enabling purging configuration files
This patch adds a parameter, EnableConfigPurge, to the base nova and
neutron service files. The value is passed to their respective puppet
modules to control whether resources not managed by puppet are purged
from the config files. The default value is set to true.

Note: this is a "semantic" backport. Instead of "service files" the
modified files in Mitaka are the overcloud, controller and compute
templates. As alluded to in the original commit message *only* the
::neutron and ::nova classes are affected. Follow up patches are
required to proliferate purging to other TripleO generated
configuration.

Conflicts (files do not exist in Mitaka):
	puppet/services/neutron-base.yaml
	puppet/services/nova-base.yaml

Change-Id: I32732dc2ed7a7b3187e8727592cb4add0c6b9cf6
Closes-Bug:  #1611800
(cherry picked from commit f7f514cf67274b40f0f079d5349d7e9fda9d1790)