Set NeutronL3HA to false when deploying DVR
This patch sets NeutronL3HA to false when the DVR environment is
included on deployment. L3 HA isn't supported with DVR routers and is
disabled by defaults in our templates but upgrading users might have it
enabled in their existing overclouds.  Including the setting here with
an appropriate comment gives us some sort of mechanism to convey to the
upgrading user what they need to do. This will have no effect on users
not using L3 HA or deploying new overclouds.

Change-Id: If036ae2b88225f5a7d5f295eb5e6874d2fbb3f72