Add NeutronGlobalPhysnetMtu to neutron-base.yaml
This adds an option to configure the global_physnet_mtu in
neutron.conf. This controls the base MTU for the physical network
which is used as a basis for the new auto calculated overlay network
MTU's.

For flat and VLAN tenant networks there has been a subtle breakage
in TripleO in that users of those network types may now need to
manually set NeutronGlobalPhysnetMtu to something less than the
physical network to imitate the previous behavior of the
now removed NeutronTenantMtu setting (removed in git
2a64b67cef74fff86ce6b56b15431b859515844d). Previously users of
say a vlan tenant network in TripleO would automatically get
the tenant network set to 1400.

Change-Id: I25c2435feb373e11e4fc3a2821db0c96211468e0