[mitaka-only] mysql: never add brackets to mysql_bind_host
The patch was submitted in Newton, but we can't backport it:
https://review.openstack.org/#/c/369369/

Don't add brackets on mysql_bind_host parameter in Galera config.
Having brackets from this parameter works with old version of
Galera but not newest one.
EPEL upgraded the version of Galera, so we need this change to deploy
TripleO in HA environment.

Change-Id: I1128c560f00fb3ffee8e072f1874ad7e610ff9e0
Closes-Bug: #1622755