Tolerate missing keys from role_data in service templates
Currently we have a few keys which may be considered optional,
such as monitoring_subscription, logging* and global_config_settings.

Currently we dereference these directly via get_attr, but this will
break when heat output validation is fixed, ref bug #1599114 is fixed
(patches are up for this, so it may be soon).

Change-Id: If4eed1ca39c10ace9b1cb5ce2dc4b9c70a3dd2f4
Partial-Bug: #1620829