Correct opencontrail hiera interpolation key
We reference keystone_auth_uri, but this isn't actually written as a
hieradata key in controller.yaml.  Intead we need to reference a
per-service key which contains the data (as these are written in
controller.yaml).  Using the neutron::server key seems reasonable since
this is enabling a neutron plugin.

Note this is a mitaka/liberty only fix because we fixed this already during
the refactoring of neutron-opencontrail.yaml into the composable services
format during the Newton cycle.

Change-Id: Id240709e256b057bbdff038aefc7fcacb0d339b8
Closes-Bug: 1638901