Mitaka-Newton upgrade fix network disruption during convergence
By default RHEL/Centos has an /etc/sysconfig/iptables file which
allows ssh and icmp only (INPUT table). During the install of
OSP9/Mitaka usually the live iptables rules are not the ones in
/etc/sysconfig/iptables but they are completely open (in ACCEPT
policy, when the user has deployed with ManageFirewall: false, which
is the default in mitaka)

Now when we run the convergence step while migrating to Newton we enable
the firewall by default and this will actually first load the rules from
/etc/sysconfig/iptables and only afterwards, it will start adding all
the rules permitting openstack traffic. This causes an outage of roughly
1 minute in our env, which disrupts the cluster and fails the
convergence step. Let's simply move the existing file out of the way, it
will be recreated by puppet in newton with the proper firewall rules
anyway.

First observed via https://bugzilla.redhat.com/show_bug.cgi?id=1391671

With this applied we saw no network connectivity loss during the
convergence step.

Closes-Bug: #1640182

Change-Id: Id4a3f7b7564a4f20e2f52a24f6d11b510ae2c4ce
Co-Authored-By: Damien Ciabrini <dciabrin@redhat.com>