pacemaker: increase timeouts for rabbitmq and redis
When we observe the 'stop timeout' values of pacemaker resources:
rabbitmq and redis, they are set to 90s. But for all other services, it
is set to 200s.
The overcloud deployment sometimes fails due to this with the error:
Error: Could not complete shutdown of rabbitmq-clone, 1 resources remaining
Error performing operation: Timer expired

This patch updates the timeout for Redis and RabbitMQ to avoid this
error.

Note: We already fixed this in puppet-tripleo in master and newton.
This commit aims at making the timeouts the same on all our releases.

Change-Id: I8a3b3951a896ee3e8e5e09778e8ea4717e76a1b4
(cherry picked from commit 76ed7fb188dbd71beb938db203d703ed7d58b5e9)