Correct AllNodesDeploySteps depends_on
This is wrong atm, it should loop to create a list for the depends_on
not multiple depends_on statements.

Note this was first corrected in https://review.openstack.org/#/c/330659/
but we need it as a standalone patch that can be backported.

Change-Id: I4d1d6346f2147e573fc0900038f1ad1d782e75ee
Closes-Bug: #1642069
(cherry picked from commit 8fe71c088889426f5a1547e96b47c569a0a0643c)