Remove conditional for neutron l3_ha
This is handled in puppet-tripleo instead so we can remove the
hard-coded reference to ControllerCount and instead use the
hiera neutron_api_node_names to derive the number of neutron API
nodes regardless of roles.

Note that the NeutronL3HA parameter is maintained despite being
marked deprecated because we need to backport this bugfix so we
can't just remove it.  I'm not sure if we want to consider removing
the deprecation as leaving the override parameter in place seems
fairly low overhead.

Closes-Bug: #1629187
Change-Id: I7a77836dcaf809cc7959fca7691a4cd7d4af5d6a
Depends-On: I01c50973eec8138ec61304f2982d5026142f267c
(cherry picked from commit 4a35893ad938b0e22767ba5c7b646ca49ccbab04)