Make the CloudDomain defaults match the doc strings
Not having the default easily accessible is causing issues for the UI,
as it cannot guess at it and can accidentally overwrite the value with
an empty string (the expected default when unset). The default is
already helpfully spelled out in the doc string for each file, this
updates the parameter to match it.

Change-Id: Ic284f9904e8f1d01cc717d59a0759f679d94106d
Closes-Bug: #1643670
(cherry picked from commit 0ca8dab4cb2acb2eb3904e4edfbbd33a47fa97a3)