Configure DPDK options to isolate PMD cores and ovs process cores
The parameter -l in DPDK_OPTIONS is applied to both list cores to be
used by PMD and the list of cores that all the OVS threads should run
on. This will impact the DPDK performance, as cores on which PMD are
running, should be completely isolated (no other process should) use
it. For this, we are configuring, -l option with list of host cpus and
the list PMD cores will be configured using ovs-vsctl config command.

Depends-On: Ib95281bcbde34c18d26ee11d4e2603b267db563b

Change-Id: Ie8e2939abfd3447b2fc028a78fd2f2e9c2d8d5b5
Fixes-Bug: #1649836