Add disable_upgrade_deployment flag to roles_data.yaml
As part of the composable upgrades current plan is to disable
the composable upgrades steps running on a particular role
(e.g. all compute nodes) in favor of a later operator driven
upgrades process as has previously been the case

This adds the disable_upgrade_deployment flag to roles_data as
a first step. Thanks to shardy for his help with this.

Change-Id: Ice845742a043b34917e61f662885786c73e955fd