Fix comment in freeipa_setup.sh script
It wrongly pointed that one can use an IP, and this is not correct. For
this case, it needs to be a CIDR. This is because we are not using DHCP
to set this address in CI, so the CIDR is required, else it will use /32
which will not route.

Change-Id: Iea070d49aeb2324564bae59082481b5485d1d360