newton/scenario001: enable ceph
I can't figure why we didn't enabled Ceph on scenario001 but it seems
reasonable to have it, it will be helpful to test upgrades of
scenario001 from newton to ocata if Ceph is already enabled.

The main difference is that we don't enable RBD backend for Gnocchi, as
we have a problem in Newton, that we never figured yet.

Change-Id: I38ba43f4047b1032e798986bca1cd5d5e74741ec