Ensure working compute node during the whole migration.
Fix subtile error introduced during the attempt to get the vm migration
after controller upgrade working.  Namely the server.cnf in the my.cnf.d
directory was updated with default values, which include listening to
127.0.0.1.  This overwrite the galera.cnf configuration and break the
link between the haproxy and the database, making the database
unavailable for services.  This happens only at convergence step where
the database is eventually restarted.

Beside fixing this bug, this is a rewrite at having the possibility for
the operator to manipulate instance during those states:
 - controllers upgraded/compute not upgraded (vm migration)
 - controllers upgraded/part of compute upgraded (vm migration)
 - controllers upgraded/all computes upgraded/no convergence

Change-Id: I22a9f1cde99c010d00c653861cc7cf34ab4cc06e
Closes-Bug: #1661202