Upgrades: wait for galera to be settled
We also need to wait for the galera resource to settle down
before we proceed starting up with the other services.

Note that before merging this, we need to land the following
change in ansible-pacemaker:
https://review.gerrithub.io/#/c/351387/

D-O is needed for upgrades to work against stable/* branches.
Depends-On: I712abe71f97c22ee3d55d9db2f641096f8a7350c

Change-Id: Id71c9cb41cfd4c17685c922db2683e28ab7588fd
Closes-Bug: #1668372
(cherry picked from commit 841d30549bd27a8b5669955196e14085025dafad)