docker-puppet.py fail if any worker fails
Currently returncodes are ignored from docker puppet workers, so a
failed puppet apply may not manifest until later in the stack
deployment due to some other failure.

This change logs any failures at the end of the run and returns a
failure code if any worker returns a failure code.

Change-Id: I6a504dbeb4c0ac465ce10e7647830524fe0a1160