Don't disable satellite repo after registration
Previously the rhel registration script disabled the satellite repo
after installing packages from it.  This means those packages will
never be updated, which is not desirable from a long-term
maintenance perspective.

I believe this behavior is a holdover from the dib registration
script, where we don't want to leave repos enabled because the
image may be deployed many times and each instance needs to be
re-registered.  In t-h-t we don't have that problem because the
script only runs at deploy time so it's okay and desirable to leave
the repos enabled.

Change-Id: I5d760467b458d90d74507a55effc49b71d22eaa3
Closes-Bug: 1673116