Increase libvirt/qemu.conf max_files and max_processes
When Nova and/or Cinder are using Ceph as backend, qemu will need to
open a connection and two threads for each and every Ceph OSD.

This change raises the max_files (set to 1024 by default) to 32768 and
the max_processes (set to 4096 by default) to 131072. The max number
of FDs is per-process, while the max number of processes is
per-user. The values can be overridden via ExtraConfig, no params are
added to the templates.

A more detailed description of the values were chosen can be found at:
https://access.redhat.com/solutions/1602683

Unclean cherry-pick from I1e79675f6aac1b0fe6cc7269550fa6bc8586e1fb

Change-Id: I1e79675f6aac1b0fe6cc7269550fa6bc8586e1fb
Depends-On: I258afd3ee6633e4b2ebc45aa8611be652476be0c
Closes-Bug: 1673995