[N->O] Fix wrong database connection for cell0 during upgrade.
During upgrade the cell0 database has the connection pointing to

   mysql+pymysql://nova:c2cdagE8PyAbnpers3AD88Hge@10.0.0.19/nova_cell0?bind_address=10.0.0.20

where 10.0.0.20 was the ip of the bootstrap node.  This makes the
nova-api fails on 2/3 node at the end of the
major-upgrade-composable-steps.yaml step.

We do have the right value in the hiera database so make sure we use
it for cell0 creation and not the nova.conf file which hasn't been
updated yet.

Change-Id: I09775206cb8fc5e15934f7e4475506a7fe17271e
Closes-Bug: #1675359