Containerize MySQL for HA
This service allows configuring and deploying MySQL/galera containers
in a HA overcloud managed by pacemaker.

The containers are managed and run by pacemaker. Inside there is
pacemaker_remote which will invoke the resource agent managing galera.
The resources themselves are created via puppet-pacemaker inside a
short-lived container used for this purpose (mysql_init_bundle).
This container needs to use the 'docker_config' section to invoke
puppet (as opposed to 'docker_puppet_tasks'), because due to the HA
composability each resource creation needs to happen on the bootstrap
node of that service and 'docker_puppet_tasks' will only run on the
controller/primary role.

Co-Authored-By: Michele Baldessari <michele@acksyn.org>
Closes-Bug: #1692842

Depends-On: I3b4d8ad2eec70080419882d5d822f78ebd3721ae

Change-Id: I790dbc30b3de1c1a3fe76d3d8f060e4d7f95e2e7