Add Docker to services list in multinode CI environments
We need the service to be present to run jobs involving containers. Note
that this is effectively a no-op for the current CI jobs, as by default
the Docker service is mapped to OS::Heat::None. Docker will actually be
deployed only if environments/docker.yaml is included in the deploy
command.

Change-Id: I97a35e30e428ff64feeb411bf63dbb7aa54f9829