Add defaults for docker puppet tasks
Currently we're referencing some steps that don't exist in the
output from the OS::Heat::Value resource, but as noted in the heat
bug #1681749 I think this isn't valid and probably should not be
allowed, so instead merge defaults with the non-empty step
tasks. To avoid further duplication of the loop variables, I
made the max step a variable.

Change-Id: Icf3d639b53c97006a0c370c12600449fba6f3323
Related-Bug: #1681749