Containerize clustercheck galera monitor for HA deployments
In HA overcloud deployments, HAProxy makes use of a helper service called
"clustercheck", to check whether galera nodes are available for serving
traffic.

This change implements a dedicated service for clustercheck, which was
originally part of the pacemaker mysql service.  The service is
configured by tripleo and the container's lifecycle is managed by docker,
like other containerized services.

Closes-Bug: #1692969
Change-Id: I8a5b30429f8ec3e484256a62a29ab7dee33ab291
Co-Authored-By: Michele Baldessari <michele@acksyn.org>
Depends-on: I1aabe34fa6a9c8c705a4405f275b66502c313cf2