Use conditionals for neutron and glance worker defaults
Using an empty string to signal that the default value in the puppet module
is to be used no longer seems to work, resulting in the puppet specified
defaults being overridden by empty string values. The impact on
configuration will differ depending on the actual configuration item, the
puppet code and the service, so it is just safer to omit the hieradata if
the user has not explicitly set a value.

Change-Id: Ic8b55941b05c4888c06e0e57b8994f71c2966034
Closes-Bug: #1669452
(cherry picked from commit Iefbc8f8669680e4f9d01db6b49543bfbe9b7661b)