Cleanup cinder paste configuration for ssl.
Newton-only.

Now that wsgi is used[1] we no longer have to configure the ssl in the
paste-api.conf file.

We remove it and revert it back to mitaka's values so that the cinder
package upgrade replace the file with the new version.

We add a basic check of the expected change to minimize the chance to
override real user change.

The sed is necessary to remove the last empty in order for the file's
md5 is the same as the original[2]

[1] https://review.openstack.org/#/q/topic:deprecate_cinder_ssl
[2] nice table of %config tag http://people.ds.cam.ac.uk/jw35/docs/rpm_config.html

Closes-Bug: #1685750

Change-Id: I53fee2314830193ecfdd8a1d92508ff855450611