[Newton only] - Manually touch ssl.conf before installing mod_ssl
In I49bd6f0017048fbab8a4011d5d7c3ffc6cda85b6 we added manual install
of mod_ssl during the mitaka to newton upgrade for the related bug
below. However we also need to prevent that package from creating
the default ssl.conf with the Listen 443 as discussed in that bug.

Related-Bug: 1682448
Depends-On: Ie6e19cc838a3f45100f6c98a350bdf6a37d40590 
Change-Id: I04b8ccf8b8637b7b4fcc0c6182ef6cd3e30a8569