Make upgrade steps unconditional to fix broken dependencies
Change I5c8b0c4abfc0607f42fd3f2da9f5ef2702b1bbe1 introduced conditions
to optimize upgrade times and fix related bugs. Unfortunately the
conditional inclusion would have to be paired with support in depends_on
to work as we need. Currently we can hit this bug if the batch upgrade
steps are undefined for some role, but upgrade steps are definied:

The specified reference "ControllerUpgradeBatch_Step2" (in
ControllerUpgradeConfig_Step0) is incorrect.

To fix this we have to make the steps unconditional. This isn't fully
reverting the original change because that change also addressed
ordering issues.

Change-Id: I369591f4757c10142f5b455e64aa778e1a9a5611
Closes-Bug: #1689553