Add sshd service to containerized compute role
This adds the sshd puppet service to the containerized compute role

All other roles already include this service from the defaults roles data,
it is only missing from the compute role.

As the sshd service runs on the docker host, this must remain as a
traditional puppet service. NB the sshd puppet service does not enable
sshd, it just enables the management of the sshd config via t-h-t/puppet.

Closes-bug: #1693837
Change-Id: I86ff749245ac791e870528ad4b410f3c1fd812e0