Containerize Cinder-backup for HA
This service allows configuring and deploying cinder-backup
containers in a HA overcloud managed by pacemaker.

The containers are managed and run by pacemaker. Pacemaker runs the
standard Kolla image but overrides the initial command so that
it explicitely calls cinder-backup. This way, we shield ourselves
from any unexpected future change in Kolla.
This container needs to use the 'docker_config' section to invoke
puppet (as opposed to 'docker_puppet_tasks'), because due to the HA
composability each resource creation needs to happen on the bootstrap
node of that service and 'docker_puppet_tasks' will only run on the
controller/primary role.

Co-Authored-By: Michele Baldessari <michele@acksyn.org>
Partial-Bug: #1668920

Depends-On: If53495ff75d4832cc6be80dc0dc9bd540ab6583b

Change-Id: Ieec823e10667592bd775bb2642f0c3790a83e85f