Containerize Cinder-volume for HA
This service allows configuring and deploying cinder-volume
containers in a HA overcloud managed by pacemaker.

The containers are managed and run by pacemaker. Pacemaker runs the
standard Kolla image but overrides the initial command so that
it explicitely calls cinder-volume. This way, we shield ourselves
from any unexpected future change in Kolla.
This container needs to use the 'docker_config' section to invoke
puppet (as opposed to 'docker_puppet_tasks'), because due to the HA
composability each resource creation needs to happen on the bootstrap
node of that service and 'docker_puppet_tasks' will only run on the
controller/primary role.

Co-Authored-By: Michele Baldessari <michele@acksyn.org>
Partial-Bug: #1668920

Depends-On: I95ad4dd89b47396bea672813d87de35e64c04b2d

Change-Id: Ib6396219c3d9484c533f6f9995d565091a197bbb