Pacemaker HA suport for OVN DB servers
This patch adds the templates required to enable the OVN DB servers
to be started in master/slave mode in the pacemaker cluster.

For the OVN DBs base profile, ::tripleo::haproxy expects the parameter
'ovn_dbs_manage_lb' set to true in order for it to configure OVN DBs
for load balancing (please see this commit [1]). So this patch sets
'ovn_dbs_manage_lb' to true.

[1] - I9dc366002ef5919339961e5deebbf8aa815c73db

Co-authored-by: Babu Shanmugam (babu.shanmugam@gmail.com)
Depends-on: I94d3960e6c5406e3af309cc8c787ac0a6c9b1756
Change-Id: I60c55abfc523973aa926d8a12ec77f198d885916
Closes-bug: #1670564