Keep floating ip reachability during pacemaker migration.
neutron-netns-cleanup-clone and neutron-ovs-cleanup-clone resources
are only one shot resources that need to be activated during bootup
and shutdown of node.  Triggering a stop has the side effect of
removing entries in the ovs db, making floating ips unreachable.

Those are triggered by the the
/usr/lib/ocf/resource.d/neutron/{OVSCleanup,NetnsCleanup} pacemaker
resource.  They, in turn, use the
/usr/lib/ocf/lib/neutron/neutron-{ovs,netns}-cleanup scripts (not the
systemd unit file).  We temporarily disable any action by configuring
the executable to be "/bin/echo" instead of the
/usr/bin/neutron-{ovs,netns}-cleanup and removing the "--force" option
in neutron-netns-cleanup.

As those resources are cloned resources we need to make sure that the
modification is done on all controller nodes before we take action on
the controller bootstraping node.  To do that we move most of Step1 to
Step0 and make the bootstrap node action happens at Step1 of the
pacemaker controller upgrade.

Furthermore we make sure that the ext bridges, if ovs is used, are not
in secure mode by setting them to standalone during the upgrade
process and back to whatever they were before.

Eventually we need to take care of the os-net-config upgrade.  It can
add new parameters to the interface definition which will force a
restart of the interfaces.  To avoid that we add the --no-activate
option.  Currently no major change in os-net-config are required for
the overcloud to continue running after the upgrade.

Co-Authored-By: Raoul Scarazzini <rscarazz@redhat.com>
Change-Id: Ib5d7b447808b51f6e436eaf6d661606132155a23
Depends-On: Ieb5ad6ad429c8388a1cbbd650339b6eecd9b7997
Closes-Bug: #1698373